+++ WavFileWriter/Downsampler.hpp
@@ -8,7 +8,7 @@ class Downsampler : public AudioStream
 {
 public:
     Downsampler();
-    void setDivider(unsigned int divider);
+    void setDivider(uint8_t divider);
     virtual void update();
 
 private:
@@ -17,8 +17,8 @@ private:
     audio_block_t*    m_inputQueueArray[1];
     audio_block_t*    m_output;
     int16_t*          m_dst;
-    unsigned int      m_samplesWritten;
-    unsigned int      m_divider;
+    uint32_t          m_samplesWritten;
+    uint8_t           m_divider;
     bool              m_firstRun;
 };
